Reorganized file structure, main screen is now returned in the Home() function. Added 3 routes to the App.tsx file
