+++ frontend/volume/src/components/login/Loading.tsx
@@ -1,0 +1,13 @@
+import "src/styles/style.css";
+
+function Loading({ LogIn }: { LogIn(): void }) {
+	return (
+		<div id="loading">
+			<div>
+				loading...
+				<img className="logo42" src="/img/42logo.svg" />
+			</div>
+		</div>
+	);
+}
+export default Loading;
